Convert token generation/validation extensions to lib
