+++ common/token-generation-lib/src/main/java/org/eclipse/dataspaceconnector/common/token/TokenGenerationService.java
@@ -1,0 +1,28 @@
+/*
+ *  Copyright (c) 2020, 2021 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Amadeus - initial API and implementation
+ *
+ */
+
+package org.eclipse.dataspaceconnector.common.token;
+
+import com.nimbusds.jwt.JWTClaimsSet;
+import org.eclipse.dataspaceconnector.spi.iam.TokenRepresentation;
+import org.eclipse.dataspaceconnector.spi.result.Result;
+import org.jetbrains.annotations.NotNull;
+
+/**
+ * Interface for generating token based on a set of claims.
+ */
+@FunctionalInterface
+public interface TokenGenerationService {
+    Result<TokenRepresentation> generate(@NotNull JWTClaimsSet claims);
+}
