+++ data-protocols/ids/ids-token-validation/src/main/java/org/eclipse/dataspaceconnector/ids/token/validation/IdsTokenValidationServiceExtension.java
@@ -13,7 +13,8 @@
  */
 
 package org.eclipse.dataspaceconnector.ids.token.validation;
-import org.eclipse.dataspaceconnector.iam.oauth2.spi.Oauth2Service;
+
+import org.eclipse.dataspaceconnector.iam.oauth2.spi.Oauth2ValidationRulesRegistry;
 import org.eclipse.dataspaceconnector.ids.token.validation.rule.IdsValidationRule;
 import org.eclipse.dataspaceconnector.spi.EdcSetting;
 import org.eclipse.dataspaceconnector.spi.system.Inject;
@@ -29,7 +30,7 @@ public final class IdsTokenValidationServiceExtension implements ServiceExtensio
     public static final String EDC_IDS_VALIDATION_REFERRINGCONNECTOR = "edc.ids.validation.referringconnector";
 
     @Inject
-    private Oauth2Service oauth2Service;
+    private Oauth2ValidationRulesRegistry oauth2ValidationRulesRegistry;
 
     @Override
     public String name() {
@@ -40,6 +41,6 @@ public final class IdsTokenValidationServiceExtension implements ServiceExtensio
     @Override
     public void initialize(ServiceExtensionContext serviceExtensionContext) {
         var validateReferring = serviceExtensionContext.getSetting(EDC_IDS_VALIDATION_REFERRINGCONNECTOR, false);
-        oauth2Service.addAdditionalValidationRule(new IdsValidationRule(validateReferring));
+        oauth2ValidationRulesRegistry.addRule(new IdsValidationRule(validateReferring));
     }
 }
