+++ extensions/data-plane-transfer/data-plane-transfer-sync/src/main/java/org/eclipse/dataspaceconnector/transfer/dataplane/sync/DataPlaneTransferSyncConfiguration.java
@@ -1,3 +1,17 @@
+/*
+ *  Copyright (c) 2022 Amadeus
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Amadeus - initial API and implementation
+ *
+ */
+
 package org.eclipse.dataspaceconnector.transfer.dataplane.sync;
 
 import org.eclipse.dataspaceconnector.spi.EdcSetting;
@@ -7,6 +21,9 @@ import java.util.concurrent.TimeUnit;
 interface DataPlaneTransferSyncConfiguration {
 
     String API_CONTEXT_ALIAS = "validation";
+
+    @EdcSetting
+    String TOKEN_SIGNER_PRIVATE_KEY_ALIAS = "edc.transfer.dataplane.token.signer.privatekey.alias";
 
     @EdcSetting
     String DATA_PLANE_PUBLIC_API_ENDPOINT = "edc.transfer.dataplane.sync.endpoint";
