+++ extensions/data-plane-transfer/data-plane-transfer-sync/src/main/java/org/eclipse/dataspaceconnector/transfer/dataplane/sync/DataPlaneTransferSyncExtension.java
@@ -14,12 +14,21 @@
 
 package org.eclipse.dataspaceconnector.transfer.dataplane.sync;
 
+import com.nimbusds.jose.JOSEException;
+import com.nimbusds.jose.JWSSigner;
+import com.nimbusds.jose.crypto.ECDSASigner;
+import com.nimbusds.jose.crypto.RSASSASigner;
+import org.eclipse.dataspaceconnector.common.token.TokenGenerationService;
+import org.eclipse.dataspaceconnector.common.token.TokenGenerationServiceImpl;
+import org.eclipse.dataspaceconnector.common.token.TokenValidationRulesRegistryImpl;
+import org.eclipse.dataspaceconnector.common.token.TokenValidationServiceImpl;
+import org.eclipse.dataspaceconnector.spi.EdcException;
 import org.eclipse.dataspaceconnector.spi.WebService;
 import org.eclipse.dataspaceconnector.spi.asset.DataAddressResolver;
 import org.eclipse.dataspaceconnector.spi.contract.negotiation.store.ContractNegotiationStore;
 import org.eclipse.dataspaceconnector.spi.iam.PublicKeyResolver;
-import org.eclipse.dataspaceconnector.spi.iam.TokenGenerationService;
 import org.eclipse.dataspaceconnector.spi.message.RemoteMessageDispatcherRegistry;
+import org.eclipse.dataspaceconnector.spi.security.PrivateKeyResolver;
 import org.eclipse.dataspaceconnector.spi.security.Vault;
 import org.eclipse.dataspaceconnector.spi.system.Inject;
 import org.eclipse.dataspaceconnector.spi.system.Provides;
@@ -30,7 +39,6 @@ import org.eclipse.dataspaceconnector.spi.transfer.flow.DataFlowManager;
 import org.eclipse.dataspaceconnector.spi.transfer.provision.ProvisionManager;
 import org.eclipse.dataspaceconnector.spi.transfer.provision.ResourceManifestGenerator;
 import org.eclipse.dataspaceconnector.spi.types.domain.transfer.StatusCheckerRegistry;
-import org.eclipse.dataspaceconnector.token.TokenValidationServiceImpl;
 import org.eclipse.dataspaceconnector.transfer.dataplane.spi.security.DataEncrypter;
 import org.eclipse.dataspaceconnector.transfer.dataplane.sync.api.controller.DataPlaneTransferSyncApiController;
 import org.eclipse.dataspaceconnector.transfer.dataplane.sync.api.resolver.SelfPublicKeyResolver;
@@ -45,22 +53,21 @@ import org.eclipse.dataspaceconnector.transfer.dataplane.sync.provision.HttpProv
 import org.eclipse.dataspaceconnector.transfer.dataplane.sync.schema.HttpProxySchema;
 import org.eclipse.dataspaceconnector.transfer.dataplane.sync.transformer.ProxyEndpointDataReferenceTransformer;
 
-import java.util.Arrays;
+import java.security.PrivateKey;
+import java.security.interfaces.ECPrivateKey;
 
 import static org.eclipse.dataspaceconnector.transfer.dataplane.sync.DataPlaneTransferSyncConfiguration.API_CONTEXT_ALIAS;
 import static org.eclipse.dataspaceconnector.transfer.dataplane.sync.DataPlaneTransferSyncConfiguration.DATA_PLANE_PUBLIC_API_ENDPOINT;
 import static org.eclipse.dataspaceconnector.transfer.dataplane.sync.DataPlaneTransferSyncConfiguration.DATA_PLANE_PUBLIC_API_TOKEN_VALIDITY_SECONDS;
 import static org.eclipse.dataspaceconnector.transfer.dataplane.sync.DataPlaneTransferSyncConfiguration.DEFAULT_DATA_PLANE_PUBLIC_API_TOKEN_VALIDITY_SECONDS;
 import static org.eclipse.dataspaceconnector.transfer.dataplane.sync.DataPlaneTransferSyncConfiguration.PUBLIC_KEY_ALIAS;
+import static org.eclipse.dataspaceconnector.transfer.dataplane.sync.DataPlaneTransferSyncConfiguration.TOKEN_SIGNER_PRIVATE_KEY_ALIAS;
 
 @Provides({EndpointDataReferenceTransformer.class})
 public class DataPlaneTransferSyncExtension implements ServiceExtension {
 
     @Inject
     private ContractNegotiationStore contractNegotiationStore;
-
-    @Inject
-    private TokenGenerationService tokenGenerationService;
 
     @Inject
     private DataAddressResolver dataAddressResolver;
@@ -96,22 +103,49 @@ public class DataPlaneTransferSyncExtension implements ServiceExtension {
      */
     @Override
     public void initialize(ServiceExtensionContext context) {
+        // common
+        var privateKeyAlias = context.getConfig().getString(TOKEN_SIGNER_PRIVATE_KEY_ALIAS);
+        var tokenSigner = createTokenSigner(context.getService(PrivateKeyResolver.class), privateKeyAlias);
+        var tokenGenerationServices = new TokenGenerationServiceImpl(tokenSigner);
+
         // api part
-        var validationRules = Arrays.asList(new ContractValidationRule(contractNegotiationStore), new ExpirationDateValidationRule());
+        var validationRulesRegistry = new TokenValidationRulesRegistryImpl();
+        validationRulesRegistry.addRule(new ContractValidationRule(contractNegotiationStore));
+        validationRulesRegistry.addRule(new ExpirationDateValidationRule());
         var resolver = createResolver(context);
-        var tokenValidationService = new TokenValidationServiceImpl(resolver, validationRules);
+        var tokenValidationService = new TokenValidationServiceImpl(resolver, validationRulesRegistry);
         webService.registerResource(API_CONTEXT_ALIAS, new DataPlaneTransferSyncApiController(context.getMonitor(), tokenValidationService, dataEncrypter));
 
         // provider part
         resourceManifestGenerator.registerProviderGenerator(new HttpProviderProxyResourceGenerator());
         statusCheckerRegistry.register(HttpProxySchema.TYPE, new HttpProviderProxyStatusChecker());
 
-        registerProvisioner(context);
+        registerProvisioner(context, tokenGenerationServices);
         dataFlowManager.register(new HttpProviderProxyDataFlowController(context.getConnectorId(), dispatcherRegistry));
         context.getTypeManager().registerTypes(HttpProviderProxyProvisionedResource.class, HttpProviderProxyResourceDefinition.class);
 
         // consumer part
-        registerTransformer(context);
+        registerTransformer(context, tokenGenerationServices);
+    }
+
+    /**
+     * Create token signer from private key.
+     */
+    private JWSSigner createTokenSigner(PrivateKeyResolver resolver, String pkAlias) {
+        var privateKey = resolver.resolvePrivateKey(pkAlias, PrivateKey.class);
+        if (privateKey == null) {
+            throw new EdcException("Failed to resolve private with alias: " + pkAlias);
+        }
+
+        if ("EC".equals(privateKey.getAlgorithm())) {
+            try {
+                return new ECDSASigner((ECPrivateKey) privateKey);
+            } catch (JOSEException e) {
+                throw new EdcException("Failed to load JWSSigner for EC private key: " + e);
+            }
+        } else {
+            return new RSASSASigner(privateKey);
+        }
     }
 
     /**
@@ -126,7 +160,7 @@ public class DataPlaneTransferSyncExtension implements ServiceExtension {
     /**
      * Register {@link EndpointDataReferenceTransformer} allowing to use consumer data plane as proxy.
      */
-    private void registerTransformer(ServiceExtensionContext context) {
+    private void registerTransformer(ServiceExtensionContext context, TokenGenerationService tokenGenerationService) {
         var endpoint = context.getConfig().getString(DATA_PLANE_PUBLIC_API_ENDPOINT);
         var consumerProxyTransformer = new ProxyEndpointDataReferenceTransformer(tokenGenerationService, dataEncrypter, endpoint, context.getTypeManager());
         context.registerService(EndpointDataReferenceTransformer.class, consumerProxyTransformer);
@@ -135,7 +169,7 @@ public class DataPlaneTransferSyncExtension implements ServiceExtension {
     /**
      * Register provider proxy provisioner serving data.
      */
-    private void registerProvisioner(ServiceExtensionContext context) {
+    private void registerProvisioner(ServiceExtensionContext context, TokenGenerationService tokenGenerationService) {
         var endpoint = context.getConfig().getString(DATA_PLANE_PUBLIC_API_ENDPOINT);
         var tokenValidity = context.getSetting(DATA_PLANE_PUBLIC_API_TOKEN_VALIDITY_SECONDS, DEFAULT_DATA_PLANE_PUBLIC_API_TOKEN_VALIDITY_SECONDS);
         var provisioner = new HttpProviderProxyProvisioner(endpoint, dataAddressResolver, dataEncrypter, tokenGenerationService, tokenValidity, context.getTypeManager());
