+++ extensions/data-plane/data-plane-api/build.gradle.kts
@@ -20,7 +20,8 @@ plugins {
 }
 
 dependencies {
-    implementation(project(":spi:web-spi"))
+    api(project(":spi:web-spi"))
+    api(project(":common:token-validation-lib"))
     implementation(project(":extensions:data-plane:data-plane-spi"))
 
     implementation("com.squareup.okhttp3:okhttp:${okHttpVersion}")
