+++ extensions/iam/oauth2/oauth2-core/build.gradle.kts
@@ -22,8 +22,7 @@ val okHttpVersion: String by project
 dependencies {
     api(project(":spi"))
     api(project(":extensions:iam:oauth2:oauth2-spi"))
-    api(project(":extensions:token:token-validation"))
-    implementation(project(":extensions:token:token-validation"))
+    implementation(project(":common:token-validation-lib"))
     implementation("com.nimbusds:nimbus-jose-jwt:${nimbusVersion}")
     implementation("com.squareup.okhttp3:okhttp:${okHttpVersion}")
