+++ extensions/iam/oauth2/oauth2-core/src/test/java/org/eclipse/dataspaceconnector/iam/oauth2/core/identity/Oauth2ServiceImplTest.java
@@ -25,8 +25,10 @@ import com.nimbusds.jose.jwk.RSAKey;
 import com.nimbusds.jose.jwk.gen.RSAKeyGenerator;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
+import org.eclipse.dataspaceconnector.common.token.TokenValidationServiceImpl;
 import org.eclipse.dataspaceconnector.iam.oauth2.core.Oauth2Configuration;
 import org.eclipse.dataspaceconnector.iam.oauth2.core.jwt.JwtDecoratorRegistryImpl;
+import org.eclipse.dataspaceconnector.iam.oauth2.core.rule.Oauth2ValidationRulesRegistryImpl;
 import org.eclipse.dataspaceconnector.spi.iam.PublicKeyResolver;
 import org.eclipse.dataspaceconnector.spi.security.CertificateResolver;
 import org.eclipse.dataspaceconnector.spi.security.PrivateKeyResolver;
@@ -57,14 +59,14 @@ class Oauth2ServiceImplTest {
 
     @BeforeEach
     void setUp() throws JOSEException {
-        RSAKey testKey = testKey();
+        var testKey = testKey();
 
         jwsSigner = new RSASSASigner(testKey.toPrivateKey());
         var publicKeyResolverMock = mock(PublicKeyResolver.class);
-        PrivateKeyResolver privateKeyResolverMock = mock(PrivateKeyResolver.class);
-        CertificateResolver certificateResolverMock = mock(CertificateResolver.class);
+        var privateKeyResolverMock = mock(PrivateKeyResolver.class);
+        var certificateResolverMock = mock(CertificateResolver.class);
         when(publicKeyResolverMock.resolveKey(anyString())).thenReturn(testKey.toPublicKey());
-        Oauth2Configuration configuration = Oauth2Configuration.Builder.newInstance()
+        var configuration = Oauth2Configuration.Builder.newInstance()
                 .tokenUrl(TOKEN_URL)
                 .clientId(CLIENT_ID)
                 .privateKeyAlias(PRIVATE_KEY_ALIAS)
@@ -75,7 +77,10 @@ class Oauth2ServiceImplTest {
                 .identityProviderKeyResolver(publicKeyResolverMock)
                 .build();
 
-        authService = new Oauth2ServiceImpl(configuration, jwsSigner, testOkHttpClient(), new JwtDecoratorRegistryImpl(), new TypeManager());
+        var validationRulesRegistry = new Oauth2ValidationRulesRegistryImpl(configuration);
+        var tokenValidationService = new TokenValidationServiceImpl(publicKeyResolverMock, validationRulesRegistry);
+
+        authService = new Oauth2ServiceImpl(configuration, jwsSigner, testOkHttpClient(), new JwtDecoratorRegistryImpl(), new TypeManager(), tokenValidationService);
     }
 
     @Test
