+++ extensions/iam/oauth2/oauth2-spi/build.gradle.kts
@@ -20,7 +20,7 @@ val nimbusVersion: String by project
 
 dependencies {
     api(project(":spi"))
-    api(project(":extensions:token:token-validation"))
+    api(project(":common:token-validation-lib"))
 
     implementation("com.nimbusds:nimbus-jose-jwt:${nimbusVersion}")
 }
