--- extensions/iam/oauth2/oauth2-spi/src/main/java/org/eclipse/dataspaceconnector/iam/oauth2/spi/Oauth2Service.java
@@ -1,25 +1,0 @@
-/*
- *  Copyright (c) 2022 Fraunhofer Institute for Software and Systems Engineering
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Fraunhofer Institute for Software and Systems Engineering
- *
- */
-
-package org.eclipse.dataspaceconnector.iam.oauth2.spi;
-
-import org.eclipse.dataspaceconnector.spi.iam.IdentityService;
-import org.eclipse.dataspaceconnector.token.JwtValidationRule;
-
-public interface Oauth2Service extends IdentityService {
-    /**
-     * Dynamically add new validation rules to the extension at runtime.
-     */
-    void addAdditionalValidationRule(JwtValidationRule additionalValidationRule);
-}
