+++ extensions/iam/oauth2/oauth2-spi/src/main/java/org/eclipse/dataspaceconnector/iam/oauth2/spi/Oauth2ValidationRulesRegistry.java
@@ -1,0 +1,9 @@
+package org.eclipse.dataspaceconnector.iam.oauth2.spi;
+
+import org.eclipse.dataspaceconnector.common.token.TokenValidationRulesRegistry;
+
+/**
+ * Registry for Oauth2 validation rules.
+ */
+public interface Oauth2ValidationRulesRegistry extends TokenValidationRulesRegistry {
+}
