--- extensions/token/token-validation/src/main/java/org/eclipse/dataspaceconnector/token/JwtValidationRule.java
@@ -1,21 +1,0 @@
-/*
- *  Copyright (c) 2022 Amadeus
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Amadeus - initial API and implementation
- *
- */
-
-package org.eclipse.dataspaceconnector.token;
-
-import com.nimbusds.jwt.SignedJWT;
-import org.eclipse.dataspaceconnector.spi.iam.ValidationRule;
-
-public interface JwtValidationRule extends ValidationRule<SignedJWT> {
-}
