+++ spi/core-spi/src/main/java/org/eclipse/dataspaceconnector/spi/types/domain/edr/EndpointDataReferenceClaimsSchema.java
@@ -22,5 +22,4 @@ package org.eclipse.dataspaceconnector.spi.types.domain.edr;
 public interface EndpointDataReferenceClaimsSchema {
     String DATA_ADDRESS_CLAIM = "dad";
     String CONTRACT_ID_CLAIM = "cid";
-    String EXPIRATION_DATE_CLAIM = "exp";
 }
