+++ system-tests/e2e-transfer-test/control-plane/build.gradle.kts
@@ -32,5 +32,5 @@ dependencies {
     api(project(":extensions:data-plane-transfer:data-plane-transfer-core"))
     api(project(":extensions:data-plane-transfer:data-plane-transfer-sync"))
     api(project(":extensions:http-receiver"))
-    api(project(":extensions:token:token-generation"))
+    api(project(":common:token-generation-lib"))
 }
