+++ system-tests/e2e-transfer-test/runner/src/test/java/org/eclipse/dataspaceconnector/test/e2e/EndToEndTransferTest.java
@@ -41,7 +41,6 @@ import java.util.concurrent.atomic.AtomicReference;
 import static io.restassured.RestAssured.given;
 import static io.restassured.http.ContentType.JSON;
 import static java.io.File.separator;
-import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 import static org.eclipse.dataspaceconnector.common.testfixtures.TestUtils.getFreePort;
@@ -66,26 +65,26 @@ public class EndToEndTransferTest {
 
     @RegisterExtension
     static EdcRuntimeExtension consumerControlPlane = new EdcRuntimeExtension(
-                ":system-tests:e2e-transfer-test:control-plane",
-                "consumer-control-plane",
-                new HashMap<>() {
-                    {
-                        put("web.http.port", String.valueOf(CONSUMER_CONTROL_PLANE.getPort()));
-                        put("web.http.path", "/api");
-                        put("web.http.validation.port", String.valueOf(CONSUMER_CONTROL_PLANE_VALIDATION.getPort()));
-                        put("web.http.validation.path", "/validation");
-                        put("edc.vault", resourceAbsolutePath("consumer-vault.properties"));
-                        put("edc.keystore", resourceAbsolutePath("certs/cert.pfx"));
-                        put("edc.keystore.password", "123456");
-                        put("edc.api.control.auth.apikey.value", API_KEY_CONTROL_AUTH);
-                        put("ids.webhook.address", CONSUMER_CONTROL_PLANE.toString());
-                        put("edc.receiver.http.endpoint", CONSUMER_BACKEND_SERVICE + "/api/service/pull");
-                        put("edc.security.private-key.alias", "1");
-                        put("edc.public.key.alias", "public-key");
-                        put("edc.transfer.dataplane.sync.endpoint", CONSUMER_DATA_PLANE_PUBLIC.toString());
-                    }
+            ":system-tests:e2e-transfer-test:control-plane",
+            "consumer-control-plane",
+            new HashMap<>() {
+                {
+                    put("web.http.port", String.valueOf(CONSUMER_CONTROL_PLANE.getPort()));
+                    put("web.http.path", "/api");
+                    put("web.http.validation.port", String.valueOf(CONSUMER_CONTROL_PLANE_VALIDATION.getPort()));
+                    put("web.http.validation.path", "/validation");
+                    put("edc.vault", resourceAbsolutePath("consumer-vault.properties"));
+                    put("edc.keystore", resourceAbsolutePath("certs/cert.pfx"));
+                    put("edc.keystore.password", "123456");
+                    put("edc.api.control.auth.apikey.value", API_KEY_CONTROL_AUTH);
+                    put("ids.webhook.address", CONSUMER_CONTROL_PLANE.toString());
+                    put("edc.receiver.http.endpoint", CONSUMER_BACKEND_SERVICE + "/api/service/pull");
+                    put("edc.transfer.dataplane.token.signer.privatekey.alias", "1");
+                    put("edc.public.key.alias", "public-key");
+                    put("edc.transfer.dataplane.sync.endpoint", CONSUMER_DATA_PLANE_PUBLIC.toString());
                 }
-        );
+            }
+    );
 
     @RegisterExtension
     static EdcRuntimeExtension consumerDataPlane = new EdcRuntimeExtension(
@@ -134,26 +133,26 @@ public class EndToEndTransferTest {
 
     @RegisterExtension
     static EdcRuntimeExtension providerControlPlane = new EdcRuntimeExtension(
-                ":system-tests:e2e-transfer-test:control-plane",
-                "provider-control-plane",
-                new HashMap<>() {
-                    {
-                        put("web.http.port", String.valueOf(PROVIDER_CONTROL_PLANE.getPort()));
-                        put("web.http.path", "/api");
-                        put("web.http.validation.port", String.valueOf(PROVIDER_CONTROL_PLANE_VALIDATION.getPort()));
-                        put("web.http.validation.path", "/validation");
-                        put("edc.vault", resourceAbsolutePath("provider-vault.properties"));
-                        put("edc.keystore", resourceAbsolutePath("certs/cert.pfx"));
-                        put("edc.keystore.password", "123456");
-                        put("edc.api.control.auth.apikey.value", API_KEY_CONTROL_AUTH);
-                        put("ids.webhook.address", PROVIDER_CONTROL_PLANE.toString());
-                        put("edc.receiver.http.endpoint", PROVIDER_BACKEND_SERVICE + "/api/service/pull");
-                        put("edc.security.private-key.alias", "1");
-                        put("edc.public.key.alias", "public-key");
-                        put("edc.transfer.dataplane.sync.endpoint", PROVIDER_DATA_PLANE_PUBLIC.toString());
-                    }
+            ":system-tests:e2e-transfer-test:control-plane",
+            "provider-control-plane",
+            new HashMap<>() {
+                {
+                    put("web.http.port", String.valueOf(PROVIDER_CONTROL_PLANE.getPort()));
+                    put("web.http.path", "/api");
+                    put("web.http.validation.port", String.valueOf(PROVIDER_CONTROL_PLANE_VALIDATION.getPort()));
+                    put("web.http.validation.path", "/validation");
+                    put("edc.vault", resourceAbsolutePath("provider-vault.properties"));
+                    put("edc.keystore", resourceAbsolutePath("certs/cert.pfx"));
+                    put("edc.keystore.password", "123456");
+                    put("edc.api.control.auth.apikey.value", API_KEY_CONTROL_AUTH);
+                    put("ids.webhook.address", PROVIDER_CONTROL_PLANE.toString());
+                    put("edc.receiver.http.endpoint", PROVIDER_BACKEND_SERVICE + "/api/service/pull");
+                    put("edc.transfer.dataplane.token.signer.privatekey.alias", "1");
+                    put("edc.public.key.alias", "public-key");
+                    put("edc.transfer.dataplane.sync.endpoint", PROVIDER_DATA_PLANE_PUBLIC.toString());
                 }
-        );
+            }
+    );
 
     @RegisterExtension
     static EdcRuntimeExtension providerBackendService = new EdcRuntimeExtension(
@@ -291,27 +290,27 @@ public class EndToEndTransferTest {
     private String createAsset(URI instance) {
         var asset = Map.of(
                 "asset", Map.of(
-                    "asset:prop:id", "asset-id",
-                    "asset:prop:name", "asset name",
-                    "asset:prop:contenttype", "text/plain",
-                    "asset:prop:policy-id", "use-eu"
+                        "asset:prop:id", "asset-id",
+                        "asset:prop:name", "asset name",
+                        "asset:prop:contenttype", "text/plain",
+                        "asset:prop:policy-id", "use-eu"
                 ),
                 "dataAddress", Map.of(
-                    "endpoint", PROVIDER_BACKEND_SERVICE + "/api/service/data",
-                    "type", "HttpData"
+                        "endpoint", PROVIDER_BACKEND_SERVICE + "/api/service/data",
+                        "type", "HttpData"
                 )
         );
 
         return given()
-            .baseUri(instance.toString())
-            .contentType(JSON)
-            .header("X-Api-Key", API_KEY_CONTROL_AUTH)
-            .body(asset)
-        .when()
-            .post("/api/assets")
-        .then()
-            .statusCode(200)
-            .extract().body().asString();
+                .baseUri(instance.toString())
+                .contentType(JSON)
+                .header("X-Api-Key", API_KEY_CONTROL_AUTH)
+                .body(asset)
+                .when()
+                .post("/api/assets")
+                .then()
+                .statusCode(200)
+                .extract().body().asString();
     }
 
     private void createContractDefinition(String assetId, URI instance) {
